Push it all down into a new TextWithDot component
