+++ cardigan/stories/components/PageHeader/PageHeader.stories.tsx
@@ -4,8 +4,7 @@ import PageHeaderStandfirst from '@weco/content/components/PageHeaderStandfirst/
 import { font } from '@weco/common/utils/classnames';
 import Picture from '@weco/common/views/components/Picture/Picture';
 import Space from '@weco/common/views/components/styled/Space';
-import Dot from '@weco/common/views/components/Dot/Dot';
-import DottedTextWrapper from '@weco/common/views/components/Dot/DottedTextWrapper';
+import TextWithDot from '@weco/common/views/components/Dot/TextWithDot';
 import PrismicImage from '@weco/common/views/components/PrismicImage/PrismicImage';
 import HeaderBackground from '@weco/common/views/components/HeaderBackground/HeaderBackground';
 import { headerBackgroundLs } from '@weco/common/utils/backgrounds';
@@ -143,20 +142,22 @@ const EventContentTypeInfo = () => (
       Saturday 8 February 2020, 13:00 – 16:00
     </Space>
     <div className="flex">
-      <DottedTextWrapper className={font('intb', 5)}>
-        <Dot dotColor="neutral.500" />
-        Past
-      </DottedTextWrapper>
+      <TextWithDot
+        className={font('intb', 5)}
+        dotColor="neutral.500"
+        text="Past"
+      />
     </div>
   </>
 );
 
 const ExhibitionContentTypeInfo = () => (
   <div className="flex">
-    <DottedTextWrapper className={font('intb', 5)}>
-      <Dot dotColor="neutral.500" />
-      Closed
-    </DottedTextWrapper>
+    <TextWithDot
+      className={font('intb', 5)}
+      dotColor="neutral.500"
+      text="Closed"
+    />
   </div>
 );
