+++ common/views/components/Dot/Dot.tsx
@@ -1,15 +1,6 @@
 import styled from 'styled-components';
 import { FunctionComponent } from 'react';
 import { PaletteColor } from '@weco/common/views/themes/config';
-import Space from '../styled/Space';
-
-const DotWrapper = styled(Space).attrs({
-  as: 'span',
-  h: { size: 'xs', properties: ['margin-right'] },
-})`
-  display: flex;
-  align-items: center;
-`;
 
 const DotEl = styled.span.attrs({
   'aria-hidden': true,
@@ -27,8 +18,6 @@ type Props = {
 };
 
 const Dot: FunctionComponent<Props> = ({ dotColor }) => (
-  <DotWrapper>
-    <DotEl dotColor={dotColor} />
-  </DotWrapper>
+  <DotEl dotColor={dotColor} />
 );
 export default Dot;
