--- common/views/components/Dot/DottedTextWrapper.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const DottedTextWrapper = styled.span`
-  display: flex;
-  align-items: center;
-`;
-
-export default DottedTextWrapper;
