+++ content/webapp/components/EventPromo/EventPromo.tsx
@@ -2,8 +2,7 @@ import { FunctionComponent } from 'react';
 import { font } from '@weco/common/utils/classnames';
 import { trackGaEvent } from '@weco/common/utils/ga';
 import LabelsList from '@weco/common/views/components/LabelsList/LabelsList';
-import Dot from '@weco/common/views/components/Dot/Dot';
-import DottedTextWrapper from '@weco/common/views/components/Dot/DottedTextWrapper';
+import TextWithDot from '@weco/common/views/components/Dot/TextWithDot';
 import EventDateRange from '../EventDateRange/EventDateRange';
 import { EventBasic } from '../../types/events';
 import { upcomingDatesFullyBooked } from '../../services/prismic/events';
@@ -162,10 +161,11 @@ const EventPromo: FunctionComponent<Props> = ({
 
           {upcomingDatesFullyBooked(event) && (
             <Space v={{ size: 'm', properties: ['margin-top'] }}>
-              <DottedTextWrapper className={font('intr', 5)}>
-                <Dot dotColor="validation.red" />
-                Fully booked
-              </DottedTextWrapper>
+              <TextWithDot
+                className={font('intr', 5)}
+                dotColor="validation.red"
+                text="Fully booked"
+              />
             </Space>
           )}
 
@@ -174,10 +174,13 @@ const EventPromo: FunctionComponent<Props> = ({
           )}
 
           {isPast && !event.availableOnline && (
-            <DottedTextWrapper className={font('intr', 5)} as="div">
-              <Dot dotColor="neutral.500" />
-              Past
-            </DottedTextWrapper>
+            <div>
+              <TextWithDot
+                className={font('intr', 5)}
+                dotColor="neutral.500"
+                text="Past"
+              />
+            </div>
           )}
         </div>
       </CardBody>
